fpashkevich: communicate between components using DI
