+++ src/app/app.component.html
@@ -1,5 +1,7 @@
-<toggle (toggled)="showMessage($event)">
-  <toggle-button></toggle-button>
+<toggle (toggled)="onToggle($event)">
   <toggle-on>On</toggle-on>
   <toggle-off>Off</toggle-off>
+  <toggle-off>Off</toggle-off>
+  <other-component></other-component>
+  <toggle-button></toggle-button>
 </toggle>
