+++ src/app/app.module.ts
@@ -3,11 +3,13 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { ToggleModule } from './toggle/toggle.module';
+import { ToggleModule } from './components/toggle/toggle.module';
+import { OtherComponent } from './components/other/other.component';
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    OtherComponent
   ],
   imports: [
     BrowserModule,
