+++ src/app/components/irrelevent-implementation-details/switch.component.html
