+++ src/app/components/irrelevent-implementation-details/switch.component.ts
