+++ src/app/components/other/other.component.html
@@ -1,0 +1,5 @@
+<div>
+  <span>Other component:</span>
+  <toggle-on>on</toggle-on>
+  <toggle-off>off</toggle-off>
+</div>
