+++ src/app/components/other/other.component.ts
@@ -1,0 +1,10 @@
+import { Component } from "@angular/core";
+
+@Component({
+  selector: 'other-component',
+  templateUrl: 'other.component.html'
+})
+
+export class OtherComponent {
+
+}
