+++ src/app/components/toggle/toggleButton/toggle.button.component.html
@@ -1,0 +1,1 @@
+<switch [on]="toggle.on" (click)="onClick()" ></switch>
