+++ src/app/components/toggle/toggleButton/toggle.button.component.ts
@@ -1,0 +1,16 @@
+import { Component } from '@angular/core';
+import { ToggleComponent } from '../toggle.component';
+
+@Component({
+  selector: 'toggle-button',
+  templateUrl: 'toggle.button.component.html'
+})
+
+export class ToggleButtonComponent  {
+
+  constructor(public toggle: ToggleComponent) {}
+
+  onClick() {
+    this.toggle.setOnState(!this.toggle.on);
+  }
+}
