+++ src/app/components/toggle/toggleOff/toggle.off.component.html
@@ -1,0 +1,1 @@
+<ng-content *ngIf="!toggle.on"></ng-content>
