+++ src/app/components/toggle/toggleOn/toggle.on.component.html
@@ -1,0 +1,1 @@
+<ng-content *ngIf="toggle.on"></ng-content>
