--- src/app/toggle/toggleButton/toggle.button.component.html
@@ -1,1 +1,0 @@
-<switch [on]="on" (click)="onClick()" ></switch>
