--- src/app/toggle/toggleOff/toggle.off.component.html
@@ -1,1 +1,0 @@
-<ng-content *ngIf="!on"></ng-content>
