--- src/app/toggle/toggleOff/toggle.off.component.ts
@@ -1,10 +1,0 @@
-import { Component, Input } from '@angular/core';
-
-@Component({
-  selector: 'toggle-off',
-  templateUrl: 'toggle.off.component.html'
-})
-
-export class ToggleOffComponent {
-  @Input() on: boolean;
-}
