--- src/app/toggle/toggleOn/toggle.on.component.html
@@ -1,1 +1,0 @@
-<ng-content *ngIf="on"></ng-content>
