--- src/app/toggle/toggleOn/toggle.on.component.ts
@@ -1,10 +1,0 @@
-import { Component, Input } from '@angular/core';
-
-@Component({
-  selector: 'toggle-on',
-  templateUrl: 'toggle.on.component.html'
-})
-
-export class ToggleOnComponent {
-  @Input() on: boolean;
-}
